add all experiments; pip_seq fails
This is because of the way experiments.config works in combination with how the test was written. The config module assumes it is loaded only once per run. However, because we load all experiments from a single test_experiments.py file, the config module is loaded for each test. Since pip_ind and pip_seq add the same config argument to the parser, the second one to do so causes a conflict.